Python based object recognition project with using OpenCV
This is a capstone level project that is must course to be graduated from school. In this project we worked within colloboratively 2 departments which is computer engineering students and electric&electronical engineering students. Project leader team was us (computer engineering students). This project aims to detect objects with camera and recognize them, after recognizition part ; objects are saved to database which is created by using SQlite3. Also, we have an Arduinos that are work for different aims. All the detailed reports are uploaded inside the file.